Create README.md
The Netflix Clone is a web application inspired by the popular streaming platform Netflix. Built entirely with HTML and CSS, this project aims to replicate the sleek design and user-friendly interface of Netflix, providing a glimpse into how modern web applications are structured and styled.

This project not only strengthened my understanding of HTML and CSS but also demonstrated my ability to design responsive web applications that prioritize user experience.